bitmask now occupies its own transfer block. moved up the stage one registers to the input in the operand matcher.

+++ accelerator/device/RTL/quartus_a10_mac/a10_mac8bitx2_c_model.cl
@@ -1,4 +1,4 @@
 #include "rtl_lib.hpp"
 int a10_mac_8bitx2 (char a0, char b0, char a1, char b1) {
-    return ( a0*b0 + a1*b1);
+    return ( ((signed int) a0)* ((signed int) b0) + ((signed int) a1)* ((signed int) b1));
 }
